# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Octobird API
  description: REST API for the Octobird GitHub Bot configuration and management.
  version: 0.1.0
  license:
    name: Apache 2.0
paths:
  /api/repos:
    get:
      summary: List registered repositories
      operationId: listRepos
      responses:
        '200':
          description: List of repository full names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "owner/repo"
  /api/repos/{owner}/{repo}/config:
    get:
      summary: Get repository configuration
      operationId: getConfig
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Repository configuration (defaults if not in DB)
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Update repository configuration
      operationId: putConfig
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '204':
          description: Configuration updated
        '400':
          description: Invalid JSON
  /api/repos/{owner}/{repo}/spam-users:
    get:
      summary: List spam users for a repository
      operationId: getSpamUsers
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of spam user accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitHubAccount'
    put:
      summary: Replace spam user list
      operationId: putSpamUsers
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/GitHubAccount'
      responses:
        '204':
          description: Spam user list updated
        '400':
          description: Invalid JSON
  /api/repos/{owner}/{repo}/mentors:
    get:
      summary: List mentors for a repository
      operationId: getMentors
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ordered list of mentor accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitHubAccount'
    put:
      summary: Replace mentor roster
      operationId: putMentors
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/GitHubAccount'
      responses:
        '204':
          description: Mentor roster updated
        '400':
          description: Invalid JSON
  /api/repos/{owner}/{repo}/audit-log:
    get:
      summary: Get recent audit log entries
      operationId: getAuditLog
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Recent audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    repoFullName:
                      type: string
                    handlerName:
                      type: string
                    action:
                      type: string
                    target:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                    details:
                      type: string
components:
  schemas:
    GitHubAccount:
      type: object
      required:
        - githubId
        - username
      properties:
        githubId:
          type: integer
          format: int64
          description: Immutable GitHub numeric user ID
          example: 12345678
        username:
          type: string
          description: Current GitHub username (may change over time)
          example: "alice"